Implemented account component functionalities

+++ src/app/services/account/account.service.ts
@@ -1,3 +1,5 @@
+import { ProductReviewService } from './../product/product-review.service';
+import { OrderService } from './../order/order.service';
 import { Account } from './../../model/account';
 import { HttpClient } from '@angular/common/http';
 import { Injectable } from '@angular/core';
@@ -13,26 +15,46 @@ export class AccountService {
     }
   }
 
-  constructor(private http: HttpClient, private cryptoService: CryptoService) { }
+  constructor(private http: HttpClient, private cryptoService: CryptoService,
+  private orderService: OrderService, private productReviewService: ProductReviewService) { }
+
+  private checkOtherServices(): Promise<boolean> {
+    return this.orderService.getTotalNumber().then(() => {
+      return this.productReviewService.getTotalNumber().then(
+        () => { return true; }, () => { return false; })
+    }, () => { return false; });
+  }
 
   public addNewAccount(newAccount: Account): Promise<Account> {
     return this.http.post<Account>
       ("http://localhost:51681/api/prodavnicaoieadmin/account/insert", newAccount, this.authOptionHeader).toPromise();
   }
 
-  public updateAccount(accountId: Number, newAccountData: Account): Promise<Account> {
+  public updateAccount(accountId: number, newAccountData: Account): Promise<Account> {
     return this.http.patch<Account>
       ("http://localhost:51681/api/prodavnicaoieadmin/account/update/" + accountId, newAccountData, this.authOptionHeader).toPromise();
   }
 
-  public deleteAccount(accountId: Number): Promise<any> {
-    return this.http.delete<any>
-      ("http://localhost:51681/api/prodavnicaoieadmin/account/delete/" + accountId, this.authOptionHeader).toPromise();
+  public deleteAccount(accountId: number): Promise<any> {
+    return this.checkOtherServices().then(response => {
+      if (response) {
+        return this.orderService.deleteAllOrdersFromAccount(accountId).then(() => {
+          return this.productReviewService.deleteAllProductReviewsByAccount(accountId).then(() => {
+            setTimeout(() => {
+              return this.http.delete<any>
+                ("http://localhost:51681/api/prodavnicaoieadmin/account/delete/" + accountId, this.authOptionHeader).toPromise()
+            }, 1000);
+          })
+        });
+      }
+      else return Promise.reject("Necessary services are not active");
+    })
+    
   }
 
-  public findAccount(accountId: Number): Promise<Account> {
+  public findAccount(accountId: number): Promise<Account> {
     return this.http.get<Account>
-      ("http://localhost:51681/api/prodavnicaoieadmin/account/find" + accountId, this.authOptionHeader).toPromise();
+      ("http://localhost:51681/api/prodavnicaoieadmin/account/find/" + accountId, this.authOptionHeader).toPromise();
   }
 
   public getTotalNumber(): Promise<Number> {
